backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
fastapi==0.135.3
google-auth==2.49.2
google-genai==1.73.1
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
lxml==6.0.4
openai==2.31.0
pillow==12.2.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.26
python-pptx==1.0.2
requests==2.33.1
sniffio==1.3.1
starlette==1.0.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
websockets==16.0
xlsxwriter==3.2.9
